correct types

+++ ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: imgwang-yeol <imgwang-yeol@student.42.f    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/09/28 21:31:32 by imgwang-yeo       #+#    #+#             */
-/*   Updated: 2020/09/28 22:29:43 by imgwang-yeo      ###   ########.fr       */
+/*   Updated: 2020/09/28 23:33:57 by imgwang-yeo      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 size_t		ft_strlcpy(char *dst, const char *src, size_t dstsize)
 {
-	int		idx;
-	int		len;
+	size_t	idx;
+	size_t	len;
 
 	len = ft_strlen(src);
 	if (dstsize == 0)
