+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: imgwang-yeol <imgwang-yeol@student.42.f    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/09/28 14:43:10 by imgwang-yeo       #+#    #+#             */
-/*   Updated: 2020/09/28 21:02:09 by imgwang-yeo      ###   ########.fr       */
+/*   Updated: 2020/09/28 23:35:27 by imgwang-yeo      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t		ft_strlen(const char *s)
 {
-	int		len;
+	size_t	len;
 
 	len = 0;
 	while (s[len])
